// src/Moryx.CommandCenter.Web/app/src/modules/models/EntryValueType.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

export enum EntryValueType {
    Byte = "Byte",
    Boolean = "Boolean",
    Int16 = "Int16",
    UInt16 = "UInt16",
    Int32 = "Int32",
    UInt32 = "UInt32",
    Int64 = "Int64",
    UInt64 = "UInt64",
    Single = "Single",
    Double = "Double",
    String = "String",
    Enum = "Enum",
    Class = "Class",
    Collection = "Collection",
}